+++ src/quick-lint-js/i18n/translation-table-generated.h
@@ -18,8 +18,8 @@ namespace quick_lint_js {
 using namespace std::literals::string_view_literals;
 
 constexpr std::uint32_t translation_table_locale_count = 5;
-constexpr std::uint16_t translation_table_mapping_table_size = 605;
-constexpr std::size_t translation_table_string_table_size = 82408;
+constexpr std::uint16_t translation_table_mapping_table_size = 606;
+constexpr std::size_t translation_table_string_table_size = 82441;
 constexpr std::size_t translation_table_locale_table_size = 35;
 
 QLJS_CONSTEVAL std::uint16_t translation_table_const_look_up(
@@ -546,6 +546,7 @@ QLJS_CONSTEVAL std::uint16_t translation_table_const_look_up(
           "switch statement is missing '{1}' around condition"sv,
           "switch statement needs parentheses around condition"sv,
           "this case will run instead"sv,
+          "this is a {1}, not a {2} ('{3}')"sv,
           "this required parameter appears after the optional parameter"sv,
           "this tuple type is a named tuple type because at least one element has a name"sv,
           "this {0} looks fishy"sv,